Pull Down Implementation
